pass extension title and name
